+++ desafio-web/src/App.test.js
@@ -1,24 +1,21 @@
 import React from "react";
 import { shallow, mount } from "enzyme";
-import TopBanner from "./components/banner/topBanner";
-import GitFeed from "./components/GitFeed";
-import { goBackToStream } from "./store/actions/goBackToStream";
-import "jest-styled-components";
 import { App } from "./App";
 import { storeFactory } from "../src/setupTests";
+import { Provider } from "react-redux";
 
 const setup = (initialState = {}) => {
   const store = storeFactory(initialState);
-  let wrapper = shallow(<GitFeed store={store} />).dive();
+  let wrapper = mount(
+    <Provider store={store}>
+      <App />
+    </Provider>
+  );
   return wrapper;
 };
 
-test("test", () => {
-  return undefined;
-});
-
-describe("testing GitFeed", () => {
-  var wrapper;
+describe("Rendering App component", () => {
+  var testComponent;
   beforeEach(() => {
     const initialState = {
       repos: [],
@@ -29,40 +26,19 @@ describe("testing GitFeed", () => {
       currentRepoOwner: null,
       currentRepo: null
     };
-    wrapper = setup(initialState);
+    testComponent = setup(initialState);
   });
 
-  test("It should exist", () => {
-    expect(wrapper).toBeTruthy;
+  it("Should initially render TopBanner and GitFeed components", () => {
+    expect(testComponent.find("topBanner")).toBeTruthy;
+    expect(testComponent.find("GitFeed")).toBeTruthy;
   });
 
-  test("it should have card components inside", () => {
-    console.log("--------", wrapper);
+  it("Should have a TopBanner with an H1", () => {
+    expect(testComponent.find("h1").length).toBe(1);
+  });
+
+  it("should not render PrFeed Component initially", () => {
+    expect(testComponent.find("PrFeed")).toBeFalsy;
   });
 });
-
-// describe("redux props", () => {
-//   test("has success piece of state as a prop", () => {
-//     const initialState = {
-//       repos: [],
-//       pageRepos: 1,
-//       pagePullRequests: 1,
-//       showPrFeed: null,
-//       pullRequests: [],
-//       currentRepoOwner: null,
-//       currentRepo: null
-//     };
-//     const success = true;
-//     const wrapper = setup(initialState);
-//     const successProp = wrapper.instance().props.success;
-//     console.log(successProp);
-//   });
-// });
-
-// describe("Action creators", () => {
-//   test("Should have an Action creator to return to repo feed", () => {
-//     expect(goBackToStream()).toEqual({
-//       type: "GO_BACK_TO_STREAM"
-//     });
-//   });
-// });
